requirements.txt
openai>=1.0.0
flask
requests
python-dotenv
plotly
pandas
numpy